requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.21
pyodbc==4.0.39
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.4
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2